+++ tests/TestCase.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Spatie\CalendarLinks\Test;
+namespace Spatie\CalendarLinks\Tests;
 
 use DateTime;
+use DateTimeImmutable;
 use DateTimeZone;
 use PHPUnit\Framework\TestCase as BaseTestCase;
 use Spatie\CalendarLinks\Link;
@@ -12,44 +13,47 @@ abstract class TestCase extends BaseTestCase
 {
     use MatchesSnapshots;
 
-    protected function createShortEventLink(): Link
+    protected function createShortEventLink(bool $immutable = false): Link
     {
-        $description = <<<'EOF'
-With balloons, clowns and stuff
-Bring a dog, bring a frog
-EOF;
+        $description = 'With balloons, clowns and stuff
+Bring a dog, bring a frog';
+
+        /** @var \DateTimeInterface $dateTimeClass */
+        $dateTimeClass = $immutable ? DateTimeImmutable::class : DateTime::class;
 
         return Link::create(
             'Birthday',
-            DateTime::createFromFormat('Y-m-d H:i', '2018-02-01 09:00', new DateTimeZone('UTC')),
-            DateTime::createFromFormat('Y-m-d H:i', '2018-02-01 18:00', new DateTimeZone('UTC'))
+            $dateTimeClass::createFromFormat('Y-m-d H:i', '2018-02-01 09:00', new DateTimeZone('UTC')),
+            $dateTimeClass::createFromFormat('Y-m-d H:i', '2018-02-01 18:00', new DateTimeZone('UTC'))
         )->description($description)->address('Party Lane 1A, 1337 Funtown');
     }
 
-    protected function createMultipleDaysEventLink(): Link
+    protected function createMultipleDaysEventLink(bool $immutable = false): Link
     {
-        $description = <<<'EOF'
-With balloons, clowns and stuff
-Bring a dog, bring a frog
-EOF;
+        $description = 'With balloons, clowns and stuff
+Bring a dog, bring a frog';
+
+        /** @var \DateTimeInterface $dateTimeClass */
+        $dateTimeClass = $immutable ? DateTimeImmutable::class : DateTime::class;
 
         return Link::create(
             'New Year',
-            DateTime::createFromFormat('Y-m-d H:i', '2019-12-31 23:00', new DateTimeZone('UTC')),
-            DateTime::createFromFormat('Y-m-d H:i', '2020-01-01 1:00', new DateTimeZone('UTC'))
+            $dateTimeClass::createFromFormat('Y-m-d H:i', '2019-12-31 23:00', new DateTimeZone('UTC')),
+            $dateTimeClass::createFromFormat('Y-m-d H:i', '2020-01-01 1:00', new DateTimeZone('UTC'))
         )->description($description)->address('Party Lane 1A, 1337 Funtown');
     }
 
-    protected function createAllDayEventLink(): Link
+    protected function createAllDayEventLink(bool $immutable = false): Link
     {
-        $description = <<<'EOF'
-With balloons, clowns and stuff
-Bring a dog, bring a frog
-EOF;
+        $description = 'With balloons, clowns and stuff
+Bring a dog, bring a frog';
+
+        /** @var \DateTimeInterface $dateTimeClass */
+        $dateTimeClass = $immutable ? DateTimeImmutable::class : DateTime::class;
 
         return Link::createAllDay(
             'Birthday',
-            DateTime::createFromFormat('Y-m-d', '2018-02-01', new DateTimeZone('UTC'))
+            $dateTimeClass::createFromFormat('Y-m-d', '2018-02-01', new DateTimeZone('UTC'))
         )->description($description)->address('Party Lane 1A, 1337 Funtown');
     }
 }